refactored / cleaned everything
