+++ src/__tests__/announcements/localizedAnnouncements.js
@@ -1,7 +1,7 @@
 // @flow
 import timemachine from "timemachine";
-import { localizeAnnouncements } from "../../announcements/logic";
-import api from "../../announcements/api/api.mock";
+import { localizeAnnouncements } from "../../providers/AnnouncementProvider/logic";
+import api from "../../providers/AnnouncementProvider/api/api.mock";
 timemachine.config({
   dateString: "February 22, 2021 13:12:59",
 });
