+++ src/providers/AnnouncementProvider/api/api.js
@@ -1,12 +1,8 @@
 // @flow
 
-import { getEnv } from "../../env";
-import network from "../../network";
-import type {
-  RawAnnouncement,
-  RawStatusSummary,
-  AnnouncementsApi,
-} from "../types";
+import { getEnv } from "../../../env";
+import network from "../../../network";
+import type { RawAnnouncement, AnnouncementsApi } from "../types";
 
 // expose a function to fetch data from the cdn (data from ledger-live-assets)
 // https://cdn.live.ledger.com/
@@ -19,19 +15,9 @@ async function fetchAnnouncements(): Promise<RawAnnouncement[]> {
 
   const { data } = await network({
     method: "GET",
-    url,
-  });
-  return data;
-}
-
-const baseStatusUrl = () => getEnv("STATUS_API_URL");
-const statusVersion = () => getEnv("STATUS_API_VERSION");
-
-async function fetchStatusSummary(): Promise<RawStatusSummary> {
-  const url = `${baseStatusUrl()}/v${statusVersion()}/summary.json`;
-
-  const { data } = await network({
-    method: "GET",
+    headers: {
+      Origin: "*",
+    },
     url,
   });
   return data;
@@ -39,7 +25,6 @@ async function fetchStatusSummary(): Promise<RawStatusSummary> {
 
 const api: AnnouncementsApi = {
   fetchAnnouncements,
-  fetchStatusSummary,
 };
 
 export default api;
