+++ src/providers/AnnouncementProvider/api/index.js
@@ -1,5 +1,6 @@
 // @flow
-import { getEnv } from "../../env";
+
+import { getEnv } from "../../../env";
 import type { AnnouncementsApi } from "../types";
 import prodApi from "./api";
 import mockApi from "./api.mock";
@@ -9,10 +10,6 @@ const api: AnnouncementsApi = {
     getEnv("MOCK")
       ? mockApi.fetchAnnouncements()
       : prodApi.fetchAnnouncements(),
-  fetchStatusSummary: () =>
-    getEnv("MOCK")
-      ? mockApi.fetchStatusSummary()
-      : prodApi.fetchStatusSummary(),
 };
 
 export default api;
