+++ src/providers/ToastProvider/types.js
@@ -1,0 +1,10 @@
+// @flow
+
+export type ToastData = {
+  id: string,
+  title: string,
+  text: string,
+  type: string,
+  icon: string,
+  callback?: () => void,
+};
